refactor PopoverMenu component
